refactor: 팝업 기능 레포지토리 분리
