+++ app/src/main/java/com/wafflestudio/snutt2/data/user/UserRepositoryImpl.kt
@@ -23,7 +23,6 @@ import com.wafflestudio.snutt2.lib.preferences.model.toDomainModel
 import com.wafflestudio.snutt2.lib.toOptional
 import com.wafflestudio.snutt2.lib.unwrap
 import com.wafflestudio.snutt2.ui.ThemeMode
-import com.wafflestudio.snutt2.views.logged_in.home.popups.PopupState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.StateFlow
 import java.util.concurrent.TimeUnit
@@ -38,7 +37,6 @@ class UserRepositoryImpl @Inject constructor(
     private val api: SNUTTRestApi,
     private val apiGoogle: SNUTTRestApiForGoogle,
     private val storage: SNUTTStorage,
-    private val popupState: PopupState,
     externalScope: CoroutineScope,
 ) : UserRepository {
 
@@ -250,54 +248,6 @@ class UserRepositoryImpl @Inject constructor(
             LoginManager.getInstance().logOut()
             storage.clearLoginScope()
             CookieManager.getInstance().removeAllCookies(null)
-            Result.Success(Unit)
-        } catch (e: Exception) {
-            Result.Fail(e.toDomainError())
-        }
-    }
-
-    override suspend fun fetchAndSetPopup(): Result<Unit> {
-        return try {
-            val popups = api._getPopup().popups.filter {
-                val expireMillis: Long? = storage.shownPopupIdsAndTimestamp.get()[it.key]
-                val currentMillis = System.currentTimeMillis()
-
-                (expireMillis == null || currentMillis >= expireMillis)
-            }
-            popupState.popup = popups
-            Result.Success(Unit)
-        } catch (e: Exception) {
-            Result.Fail(e.toDomainError())
-        }
-    }
-
-    override suspend fun closePopupWithHiddenDays(): Result<Unit> {
-        return try {
-            val popup = popupState.popup.firstOrNull()
-            if (popup != null) {
-                val expiredDay: Long = popup.popupHideDays?.let { hideDays ->
-                    System.currentTimeMillis() + TimeUnit.DAYS.toMillis(hideDays.toLong())
-                } ?: INFINITE_LONG_MILLIS
-
-                storage.shownPopupIdsAndTimestamp.update(
-                    storage.shownPopupIdsAndTimestamp.get()
-                        .toMutableMap()
-                        .also {
-                            it[popup.key] = expiredDay
-                        },
-                )
-
-                popupState.popup = popupState.popup.drop(1)
-            }
-            Result.Success(Unit)
-        } catch (e: Exception) {
-            Result.Fail(e.toDomainError())
-        }
-    }
-
-    override suspend fun closePopup(): Result<Unit> {
-        return try {
-            popupState.popup = popupState.popup.drop(1)
             Result.Success(Unit)
         } catch (e: Exception) {
             Result.Fail(e.toDomainError())
